Add more generic functions.
