+++ deploy/CAD_Installs/packages.config
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <!-- n.b the last component of the versions is ignored, and replaced with the last svn modification revision -->
 <packages>
-  <package id="META.CadCreoParametricCreateAssembly" version="1.5.23.0" />
+  <package id="META.CadCreoParametricCreateAssembly" version="1.5.24.0" />
   
   <package id="META.ExtractACM-XMLfromCreoModels" version="1.5.2.0" />
