+++ src/CADAssembler/CADCommonFunctions/cc_AssemblyUtilities.h
@@ -241,6 +241,71 @@ namespace isis
 									std::map<std::string, isis::CADComponentData> &in_CADComponentData_map )
 												throw (isis::application_exception);
 
+
+
+
+	////////////////////////////////////////////////////////////////////////////////////////////////////////
+	//	Description:
+	//		This function adds ComponentInstanceIDs to the 
+	//		in_out_CADComponentData_map::…::constrainedTo_ComponentInstanceIDs_DerivedFromConstraintPairs, wherein the 
+	//		ComponentInstanceIDs are either the part IDs that the datums contained in in_out_CADComponentData_map::…::constraintPairs
+	//		reference, or the assembly IDs for the assemblies that are treated as one body (i.e. CADAssembly.xml 
+	//		special instruction != HAS_KINEMATIC_JOINT ). Reference means the parts that contain the geometry 
+	//		referenced by a datum.  See the contract for 
+	//		FindPartsReferencedByFeature for more information about the meaning of a referenced part.
+	//
+	//	Pre-Conditions:
+	//		in_TopAssemblyComponentInstanceID must be the top assembly for the entire assembly.  The feature IDs leading 
+	//		to a sub-part/sub-assembly start at the top assembly.
+	//
+	//		in_AssemblyComponentIDs must be ComponentIDs in in_out_CADComponentData_map. 
+	//
+	//	Post-Conditions:
+	//		if in_AssemblyComponentIDs.size() == 0, no action taken
+	//
+	//		if a referenced ComponentID is not found in in_FeatureIDs_to_ComponentInstanceID_hashtable
+	//			then
+	//				throw isis::application_exception
+	void PopulateMap_with_ConstrainedToInfo_per_InputXMLConstraints ( 
+			cad::CadFactoryAbstract													&in_Factory,
+			const std::string														&in_TopAssemblyComponentInstanceID,
+			const std::vector<std::string>											&in_AssemblyComponentIDs,
+			const std::unordered_map<IntList, std::string, ContainerHash<IntList>>	&in_FeatureIDs_to_ComponentInstanceID_hashtable,
+			std::map<std::string, isis::CADComponentData>							&in_out_CADComponentData_map )
+																		throw (isis::application_exception);
+
+
+
+	//	Description:
+	//		Returns if in_ComponentInstanceID is a CyPhy leaf assembly.  This is  determineed  by checking if 
+	//		the immediate (i.e. not grandchildren) children of in_ComponentInstanceID have the setting of 
+	//		INITIAL_SOURCE_DERIVED_FROM_LEAF_ASSEMBLY_DESCENDANTS
+	//
+	//	Pre-Conditions:
+	//		forEachLeafAssemblyInTheInputXML_AddInformationAboutSubordinates must have been invoked before 
+	//		calling this function; and if not, CyPhyLeafAssembly will always return false.
+	//
+	//	Post-Conditions:
+	//		if in_ComponentInstanceID is a part
+	//			return false
+	//		if the first immediate child has the setting of INITIAL_SOURCE_DERIVED_FROM_LEAF_ASSEMBLY_DESCENDANTS
+	//			return true
+	//		else
+	//			return false
+	bool CyPhyLeafAssembly (	const std::string								&in_ComponentInstanceID,
+							std::map<std::string, isis::CADComponentData>	&in_out_CADComponentData_map );
+
+
+	void AddBoundingBoxValuesToMap( cad::CadFactoryAbstract				&in_Factory,
+						const std::vector<std::string>					&in_AssemblyComponentIDs,
+						std::map<std::string, isis::CADComponentData>	&in_CADComponentData_map) throw (isis::application_exception);
+
+	void AddMassPropertyValuesToMap( 
+						cad::CadFactoryAbstract							&in_Factory,
+						const std::vector<std::string>					&in_AssemblyComponentIDs,
+						std::map<std::string, isis::CADComponentData>	&in_out_CADComponentData_map) 
+																				throw (isis::application_exception);
+
 }
 
 #endif
