+++ src/CADAssembler/CADCreoParametricCreateAssembly/ISISVersionNumber.h
@@ -1,7 +1,7 @@
 // The rc compiler will not except #ifndef blocks.  Therefore, include this file judiciously.
-#define ISIS_PRODUCT_VERSION  1,5,23,0
+#define ISIS_PRODUCT_VERSION  1,5,24,0
 #define ISIS_FILE_VERSION ISIS_PRODUCT_VERSION
-#define ISIS_STR_PRODUCT_VERSION "CADCreoParametricCreateAssembly 1.5.23.0\0"
-#define ISIS_PRODUCT_VERSION_WITH_v_AND_DOTS "v1.5.23.0"
+#define ISIS_STR_PRODUCT_VERSION "CADCreoParametricCreateAssembly 1.5.24.0\0"
+#define ISIS_PRODUCT_VERSION_WITH_v_AND_DOTS "v1.5.24.0"
 #define ISIS_CYPHY_2_CAD_DLL_MIN_VERSION "1.2.0.0"
 #define ISIS_METRIC_FILE_VERSION "v1.1.0.0"
